Add m_lusersnoservices and m_rpg modules.
Both upgraded from my 1.2 versions to 2.0.

The m_lusersnoservices module adds a line to LUSERS that displays the global count without Services (assuming that Services are the only U-Lines on the network).

The m_rpg module adds the commands NPC, NPCA, AMBIANCE, NARRATOR and NARRATORA for use in pen and paper RPGs being played over IRC. For the 2.0 update, it has also been made to actually work on a network by broadcasting with ENCAP.