Add operoverride and reply on block abilities
Upon request the following has been added:
* Oper override ability using the 'users/override-blockinvite' priv.
* A config option to block silently or reply to source when
the INVITE is blocked.

While updating this, it was found that checking IS_LOCAL on the 'dest'
user was incorrect. This worked fine without user priv checking and not
writing a numeric back to the source. With the additions, we need to
only act on a local source user (which is the correct method anyways).