Fixed all review requests + small rank check
Fixed all Review requests. The delrank was a tricky one as the innerban map had to be put back into the outer map when the element of the inner map has been deleted.

Also I noticed that mh->IsListMode() returns true also for OP/Half-OP/Admin/Founder prefix modes (seems like these internally is implemented as list modes) so I added a check that prevents the logic to be run if the mode is a prefix mode.