[SECURITY FIX] Added check to prevent targeting of other oper's
Added check to prevent /PRETENDUSER on Another oper, which would of course bypass any access restrictions that the executing oper has. Of course, /PRETENDUSER is a pretty serious command and should Only be given to trusted opers, but still, I think it shouldn't be possible to execute this command on Another network operator.

Imagine you dont have rights to /LOADMODULE, but have /PRETENDUSER, but you know the nick of an online oper that has the right.
And you just run /PRETENDUSER [that_oper] LOADMODULE m_restrictchans.so

Can cause pretty serious distruptions in a network.