Add module for snomasks p and P (User PMs)
I got a request for creating this module from a server owner on Another network via email, since this owner had large problems with abuse over PMs that fly under the radar.
This module allows server administrators (and opers) to read PMs through Snomasks p and P. (p = local PMs, P = remote PMs).
This allows server administrators to enforce network usage rules across PMs.
Note that on larger networks this module might get spammy, why I recommend instead redirecting the output to a channel like #userpm using m_chanlog.so, and don't forget +O on that channel.

Config:
<userpm ignoreuline="yes/no" ignoreoper="yes/no" expandnicks="0/1/2">

ignoreuline="yes"/"no": This will make the module ignore PMs to/from u-lined servers if set. This config is HIGHLY RECOMMENDED to set to "yes" to avoid NickServ/ChanServ/OperServ passwords from ending up in the wrong places. However, for some networks it might be useful to set this to "no" to make it possible to log excessive login attempts to NickServ/ChanServ/OperServ and use a bot to issue Z-lines/G-lines/etc.

ignoreoper="yes"/"no": This will make the module ignore PMs to/from opers. In some cases it can be a good idea to set it to "yes" to avoid PMs to/from Opers from ending up as server notices, but in some cases it can be useful to see other oper's PMs (for example if an oper does an decision in favor for a user, for example "Yes you can bring your own bot here", it could be useful for a network at a large to be able to see this decision and prevent other oper's from intervening thinking the user did not have permission)

expandnicks="0/1/2":
This config allows you to set how nicks are displayed in the server notices. If you set this to 0, only the nick is shown, for example:
USERPM: <Source>,<Target>: Content of PM

If you set it to 1, it will display the nicks in expanded nick!ident@host format, but will leave host cloaked, if you have the cloak module activated, for example:
USERPM: <Source!Source@network-gs7.9vf.8gbi23.IP>,<Target!Target@network-h8j.2dv.2nbjkl.IP>: Content of PM

If you set it to 2, it will display it in uncloaked format, for example:
USERPM: <Source!Source@user1.someisp.com>,<Target!Target@user2.anotherisp.com>: Content of PM

The module displays everything in a easy-to-parseable format so its also possible to create bots that act on PM content.
To disable the possibility to use OTR encryption, use a censor/filter module to disable the string "?OTR" to be sent.

Note that this module can be a Little controversial in the IRC World, why I recommend to be responsible when using this module on your server, and not allowing every oper to use this module. More about restricting usage to certain opers later down.

To restrict which opers that are able to read PMs through snomasks, I recommend disabling mode "s" in server config (so no opers can use SNOMASKS), and instead provide SNOMASK functionality through separate +O channels using m_chanlog.so, where you can through channel config restrict which opers have access (for example via bans, exceptions, invites and channel keys), which allows diverse access condition for each snomask.