Add m_restrictmsg_duration - Restrict messages for a specified duration (#163)
Restrict messages until a user has been connected for a specified duration.
This is another spam fighting method. Exemptions are included for messages
to opers, to U-Lined clients, and from registered and identifed users.
Notifying of the restriction is configurable and the events are
prioritized last so things like FILTER, +R, bans, etc. act first.