Bugfix: XCOPY and XREMOVE: not checking command permissions
- Currently any oper with the XCOPY command permission can copy
an X-Line of any type, even if their oper class doesn't allow them
to set that X-Line by way of the normal command.
- Likewise any oper with the XREMOVE command permission can remove
an X-Line of any type.
- Change both commands over to just local handling so we can check
LocalUser::HasPermission when necessary. Return a 'Permission Denied'
numeric (or with '*' removal, a notice) when they are missing the
appropriate command permission.
- Also wrap all the common functions in an anonymous namespace.

Thanks to @Robby- for spotting and reporting this bug.