Improve CIDR mask matching.
Since MatchCIDR() is meant to take a user's IP address and a CIDR mask,
having a CIDR mask as the first argument (the X-line mask) does not work.
To work-around this for now, run the function 'both ways' to match a given
CIDR mask to an address or a given address to a CIDR mask.
This isn't the best solution as it won't work if both the X-line mask and
the given mask are CIDR masks; I'll work on a solution for that.