Make hack in m_hash_crypt a lot less hacky.
Basically, we know that linux, freebsd, and netbsd require -lcrypt.
openbsd and darwin don't (and the build will fail on at least macos if
we add that flag). So, add the known systems to the LinkerFlags and go
on our merry way.

The reason we can't use pkg-config is because libcrypt-dev is not widely
available yet (Debian stale notably does not have it) and operating
systems like FreeBSD are unlikely to really ever have it.