Use ephemeral fake user for roleplay clients
This is in a simlar vein to FakeUser, but inherits directly from User
since this isn't a server (although it pretends to be one, it's
ultimately just... us). This is good enough to fool most modules like
the history module and spanningtree module.

This fixes a bug where spanningtree would erroneously propagate roleplay
messages as if they came from the person who used the command. This also
happens to fix an issue where the chanhistory module would not save the
logs correctly from roleplay.